+++ android/app/src/main/java/com/foss/foss/feature/matchsearching/recent/RecentMatchesEvent.kt
@@ -1,4 +1,4 @@
-package com.foss.foss.feature.matchearching.recent
+package com.foss.foss.feature.matchsearching.recent
 
 sealed interface RecentMatchesEvent {
