+++ android/app/src/main/res/layout/fragment_recent_matches.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#171B1E"
-    tools:context=".feature.matchearching.recent.RecentMatchesFragment">
+    tools:context=".feature.matchsearching.recent.RecentMatchesFragment">
 
     <TextView
         android:id="@+id/recent_tv_info"
