+++ android/app/src/main/res/layout/fragment_relative_matches_detail.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#171B1E"
-    tools:context=".feature.matchearching.recent.RecentMatchesFragment">
+    tools:context=".feature.matchsearching.recent.RecentMatchesFragment">
 
     <TextView
         android:id="@+id/relative_detail_view_tool_bar"
