+++ android/app/src/test/java/RelativeMatchesViewModelTest.kt
@@ -1,6 +1,6 @@
 import app.cash.turbine.test
-import com.foss.foss.feature.matchearching.relative.RelativeMatchesEvent
-import com.foss.foss.feature.matchearching.relative.RelativeMatchesViewModel
+import com.foss.foss.feature.matchsearching.relative.RelativeMatchesEvent
+import com.foss.foss.feature.matchsearching.relative.RelativeMatchesViewModel
 import com.foss.foss.model.Match
 import com.foss.foss.model.MatchMapper.toUiModel
 import com.foss.foss.model.RelativeMatch
